ALIA: Instantiate select over stores
