+++ src/logics/CMakeLists.txt
@@ -11,6 +11,7 @@ target_sources(logics
 	PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/Logic.h"
 	PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/LATheory.h"
 	PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/ArrayTheory.cc"
+	PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/ArrayHelpers.cc"
 	PUBLIC "${CMAKE_CURRENT_SOURCE_DIR}/ArithLogic.h"
 	PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/ArithLogic.cc"
 	PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/Theory.cc"
